Curriculum: Version 2 of Learn Clojure with Shapes
This major iteration of our introduction to programming consists of a
significant rewrite plus incorporating lessons from field-testing it
at ClojureBridge Berlin 2017. (Many of those lessons were tracked in
the checklist on the Curriculum page of the wiki.)

Some major points:

 - hammer repeatedly at the concept of a Code Inspection Toolbox --> we
   want students to have something to reach for when they feel confused
 - draw a shape first
 - use simpler and more fun examples for functions
 - less talk, more do when it comes to `map`
 - more variety in size, shape, and color of example sketches
 - nix the dogfooded fn/defn diagrams in favor of an image
 - introduce a structured "solution template"

Unfortunately the length (measured in screen-heights) of the intro
increased about 20%. We would like to shorten it if we think that is
possible without cutting into muscle.

This curriculum update is related to recent commits that update the
Editor Quickstart.