Errors: fix bug in trie search
Previously, the context of each step while searching the trie was
calculated independently of the check for the next match. This causes
problems when a context match is made on the current token, but the
next token matches on a different branch than the context was
found. For example, if our tokens are:

[the function foo is broken at bar]

then we'd like our context at the final recursive step to be `[foo bar]`.

Assume the keys of the top level of the trie are:

[%, the, oops]

then on our first iteration we will match on the wildcard "%" as well
as "the" itself. On the next recursive step we will match only on
"function". However, if we calculate `context` based only on our
current knowledge at this step, we will put "the" into the vector of
context matches, because it matches the wildcard sibling to our token
match. This is incorrect. We only want the current token to be added
to the `context` if it is part of the trie branch we are following.

Therefore, this commit alters the trie search logic such that the
current token is only added to the context when we have a match on the
wildcard, not the current token.