Errors: don't split tokens on `_`
This allows us to handle "X is not a function" errors much better,
because JS-ified function names like "a_b_c_fn" are kept as a single
String while matching, and the underscores are replaced by dashes
afterwards for display to the user.

This commit also splits error messages at the line break, for easier reading.