ClojureBridge London curriculum
We have been using Maria.cloud for ClojureBridge London since last year and
evolved the curriculum to include the most valuable aspects from several
curriculums in one place and add some more examples.

Would this be a welcome addition to the curriculum's already available via
Maria.cloud directly?

Currently we ask students to fork a this proposed curriculum from a Gist to
their own GitHub account, where they can then access that Gist once they login
to Maria.cloud.

https://clojurebridgelondon.github.io/workshop/colours-and-shapes/

It would be a simpler process if the ClojureBridgeLondon curriculum was included
in Maria.cloud directly.  However, I appreciate you have a particular style you
wish to maintain, so if it does not fit along with the existing curriculum's
then we understand.